Test the paths

+++ scripts/download-externals.ps1
@@ -3,6 +3,8 @@ param (
     [string] $Artifact
 )
 
+$ErrorActionPreference = 'Stop'
+
 Add-Type -AssemblyName System.IO.Compression.FileSystem
 
 $url = "https://dev.azure.com/xamarin/public/_apis/build/builds/$BuildNumber/artifacts?artifactName=$Artifact&api-version=5.1&%24format=zip"
@@ -13,11 +15,13 @@ $output = "./output"
 
 # dowload the artifact
 New-Item -Type Directory -Path $outputTemp -Force | Out-Null
-# Invoke-WebRequest -Uri $url -OutFile $dest
+Invoke-WebRequest -Uri $url -OutFile $dest
 
 # extract it
-Remove-Item -Path $destTemp/** -Force -Recurse
-[System.IO.Compression.ZipFile]::ExtractToDirectory($dest, $outputTemp)
+if (Test-Path -Path $destTemp) {
+    Remove-Item -Path $destTemp/** -Force -Recurse
+    [System.IO.Compression.ZipFile]::ExtractToDirectory($dest, $outputTemp)
+}
 
 # create the output folder
 New-Item -Type Directory -Path $output -Force | Out-Null
